user defined repository functions tests
