+++ delivery-engine/src/test/java/ua/tqs21/deliveryengine/auth/AuthorizationTest.java
@@ -1,4 +1,4 @@
-package ua.deliveryengine.auth;
+package ua.tqs21.deliveryengine.auth;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
