+++ delivery-engine/src/test/java/ua/tqs21/deliveryengine/controller/RiderControllerTest.java
@@ -1,21 +1,29 @@
-package ua.deliveryengine.controller;
+package ua.tqs21.deliveryengine.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.mockito.Mock;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.MediaType;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.servlet.MockMvc;
 import ua.tqs21.deliveryengine.DeliveryEngineApplication;
+import ua.tqs21.deliveryengine.dto.RiderPostDTO;
 import ua.tqs21.deliveryengine.enums.Roles;
+import ua.tqs21.deliveryengine.models.Order;
+import ua.tqs21.deliveryengine.models.Rider;
 import ua.tqs21.deliveryengine.models.User;
-import ua.tqs21.deliveryengine.repositories.UserRepository;
+import ua.tqs21.deliveryengine.repositories.RiderRepository;
+
+import java.util.Date;
+import java.util.HashSet;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.hamcrest.CoreMatchers.is;
@@ -27,34 +35,43 @@ import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 
-
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.MOCK, classes = DeliveryEngineApplication.class)
 @AutoConfigureMockMvc
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
 @ActiveProfiles("test")
-public class UserControllerTest {
+public class RiderControllerTest {
 
     @Autowired
     private MockMvc mvc;
 
-    @Autowired
-    private UserRepository userRepository;
+    @Mock
+    private PasswordEncoder passwordEncoder;
 
-    private User user1 = new User("base1", "psw", Roles.ADMIN.name());
-    private User user2 = new User("base2", "psw", Roles.RIDER.name());
+    @Autowired
+    private RiderRepository riderRepository;
+
+
+    private User base1 = new User("base1", "psw", Roles.RIDER.name());
+    private User base2 = new User("base2", "psw", Roles.RIDER.name());
+    private Rider rider1 = new Rider(base1, new HashSet<Order>());
+    private Rider rider2 = new Rider(base2, new HashSet<Order>());
+
+
     @BeforeEach
     void setUp(){
-        userRepository.saveAndFlush(user1);
-        userRepository.saveAndFlush(user2);
+        riderRepository.saveAndFlush(rider1);
+        riderRepository.saveAndFlush(rider2);
     }
 
     @AfterEach
     void cleanUp(){
-        userRepository.deleteAll();
+        riderRepository.deleteAll();
     }
+
     @Test
-    void whenGetUser_thenGetAllUsers() throws Exception {
-        mvc.perform(get("/user/")
+    @WithMockUser
+    void whenGetRider_thenGetAllRiders() throws Exception {
+        mvc.perform(get("/courier/")
                         .contentType(MediaType.APPLICATION_JSON))
                 .andDo(print())
                 .andExpect(content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON))
@@ -62,36 +79,33 @@ public class UserControllerTest {
     }
 
     @Test
-    void whenPostUser_userIsAddedToRepository() throws Exception{
-        mvc.perform(get("/user/")
-                .contentType(MediaType.APPLICATION_JSON));
-        assertThat(userRepository.findAll().size()).isEqualTo(2);
+    @WithMockUser
+    void whenPostRider_riderIsAddedToRepository() throws Exception{
+        RiderPostDTO rider3 = new RiderPostDTO("base3", "http://teste.com/a.jpg", new Date(), "psw3");
+        mvc.perform(post("/courier/")
+                .content(asJsonString(rider3))
+                .contentType(MediaType.APPLICATION_JSON))
+                .andExpect(jsonPath("$.user.email", is(rider3.getEmail())));
+        assertThat(riderRepository.findAll().size()).isEqualTo(3);
     }
 
     @Test
     @WithMockUser
-    void whenFindUserById_findUser() throws Exception{
-        mvc.perform(get("/user/"+user1.getId()))
+    void whenFindRiderById_findRider() throws Exception{
+        mvc.perform(get("/courier/"+rider1.getId()))
                 .andDo(print())
-                .andExpect(jsonPath("$.id", is(user1.getId())))
-                .andExpect(jsonPath("$.email", is(user1.getEmail())));
+                .andExpect(jsonPath("$.id", is(rider1.getId())))
+                .andExpect(jsonPath("$.user.id", is(rider1.getUser().getId())));
     }
 
     @Test
-    void whenUpdateUser_updateUser() throws Exception{
-        mvc.perform(put("/user/")
-                .content(asJsonString(user1))
-                .contentType(MediaType.APPLICATION_JSON))
-                .andExpect(jsonPath("$.id", is(user1.getId())))
-                .andExpect(jsonPath("$.email", is(user1.getEmail())));
-    }
-
-    @Test
-    void whenDeleteUserById_deleteUser() throws Exception{
-        mvc.perform(delete("/user/"+user1.getId()))
+    @WithMockUser
+    void whenDeleteRiderById_deleteRider() throws Exception{
+        int id = rider1.getId();
+        mvc.perform(delete("/courier/"+id))
                 .andDo(print())
-                .andExpect(jsonPath("$", is(user1.getId())));
-        assertThat(userRepository.findAll().size()).isEqualTo(1);
+                .andExpect(jsonPath("$", is(id)));
+        assertThat(riderRepository.findAll().size()).isEqualTo(1);
     }
 
     public static String asJsonString(final Object obj) {
