+++ delivery-engine/src/test/java/ua/tqs21/deliveryengine/models/DataSchemaTest.java
@@ -1,4 +1,4 @@
-package ua.deliveryengine.models;
+package ua.tqs21.deliveryengine.models;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
