+++ delivery-engine/src/test/java/ua/tqs21/deliveryengine/utils/AddressResolverTest.java
@@ -1,4 +1,4 @@
-package ua.deliveryengine.utils;
+package ua.tqs21.deliveryengine.utils;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
